# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample API sandbox for testing node API development.",
    "version": "1.0.0",
    "title": "Sandbox API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "user",
      "description": "Operations for users"
    },
    {
      "name": "tasks",
      "description": "TODO-style task operations for users"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/authenticate": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Authenticate a user",
        "operationId": "authenticate",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "authenticationRequest",
            "schema": {
              "type": "object",
              "required": [
                "username"
              ],
              "properties": {
                "username": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Invalid user"
          }
        }
      }
    },
    "/user/{userId}/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Finds all tasks for a single user",
        "operationId": "allUserTasks",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create a new task for the user",
        "operationId": "createUserTask",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "detail",
            "schema": {
              "$ref": "#/definitions/TaskDetail"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/tasks/{taskId}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Finds a single task",
        "operationId": "getUserTask",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Update a single task",
        "operationId": "updateUserTask",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "detail",
            "schema": {
              "$ref": "#/definitions/TaskDetail"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete a single task",
        "operationId": "deleteUserTask",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "profile": {
          "$ref": "#/definitions/UserProfile"
        },
        "created": {
          "$ref": "#/definitions/DateTime"
        },
        "modified": {
          "$ref": "#/definitions/DateTime"
        }
      }
    },
    "UserProfile": {
      "type": "object",
      "properties": {
        "last": {
          "type": "string"
        },
        "first": {
          "type": "string"
        }
      }
    },
    "DateTime": {
      "type": "string",
      "format": "YYYY-MM-DD HH:MM:SS.######",
      "example": "2019-01-02 20:39:23.709599"
    },
    "Task": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "user_id": {
          "type": "integer",
          "format": "int64"
        },
        "detail": {
          "$ref": "#/definitions/TaskDetail"
        },
        "created": {
          "$ref": "#/definitions/DateTime"
        },
        "modified": {
          "$ref": "#/definitions/DateTime"
        }
      }
    },
    "TaskDetail": {
      "type": "object",
      "properties": {
        "complete": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        }
      }
    }
  }
}